requirements.txt
cachetools==2.0.0
certifi==2017.4.17
chardet==3.0.4
dill==0.2.6
Django==1.11.2
entrypoints==0.3
flake8==3.7.8
future==0.16.0
gapic-google-cloud-speech-v1==0.15.3
google-auth==1.0.1
google-auth-httplib2==0.0.2
google-cloud-core==0.24.1
google-cloud-speech==0.25.1
google-gax==0.15.13
googleapis-common-protos==1.5.2
grpcio==1.3.5
gunicorn==19.7.1
httplib2==0.10.3
idna==2.5
isort==4.3.21
mccabe==0.6.1
oauth2client==3.0.0
pafy==0.5.3.1
ply==3.8
proto-google-cloud-speech-v1==0.15.3
protobuf==3.3.0
psycopg2==2.7.1
pyasn1==0.2.3
pyasn1-modules==0.0.9
pycodestyle==2.5.0
pydub==0.23.1
pyflakes==2.1.1
pytz==2017.2
requests==2.18.1
rsa==3.4.2
six==1.10.0
urllib3==1.21.1
youtube-dl==2019.10.22
